% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowToCol.R
\name{RowToCol}
\alias{RowToCol}
\title{Row to Column}
\usage{
RowToCol(m)
}
\arguments{
\item{m}{The name of the matrix that will be processed}
}
\description{
This function transforms a single column Matrix into a single column
}
\examples{
RowToCol()
}
